+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 #########
 
+0.1.19 - Under development
+--------------------------
+
+* Add a ``othername`` field on on the ``Person`` model
+* Add a ``letter_prefix`` property on the ``Person`` model
+
 0.1.18
 ------